requirements.txt
PyMySQL==1.1.0
cryptography==41.0.7
tk==0.1.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
Flask==3.0.0
requests>=2.31.0